Add method to unzip Data to a FileWrapper
The existing unzip method takes a URL and returns the filename of the unzipped data on disk. This is not terribly convenient.

Instead, we create a helper method that takes in Data and writes it to the cache directory, piping that file’s URL into the existing method. It then takes the returned result from the existing method and loads it into a FileWrapper, and returns that.

This makes it easier to work with zipped NSDocument bundles.